refactor(test): 테스트 패키지 분리
- 테스트 코드도 패키지를 분리하여 관리하도록 변경하였습니다.
