+++ src/test/java/kr/hhplus/be/server/coupon/domain/entity/UserCouponStateTest.java
@@ -1,9 +1,8 @@
-package kr.hhplus.be.server.coupon;
+package kr.hhplus.be.server.coupon.domain.entity;
 
 import kr.hhplus.be.server.common.BusinessException;
 import kr.hhplus.be.server.common.ErrorCode;
 import kr.hhplus.be.server.coupon.domain.UserCouponStatus;
-import kr.hhplus.be.server.coupon.domain.entity.UserCouponState;
 import kr.hhplus.be.server.coupon.fixture.UserCouponStateFixture;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
