+++ src/test/java/kr/hhplus/be/server/point/domain/entity/UserPointTest.java
@@ -1,8 +1,7 @@
-package kr.hhplus.be.server.point;
+package kr.hhplus.be.server.point.domain.entity;
 
 import kr.hhplus.be.server.common.BusinessException;
 import kr.hhplus.be.server.common.ErrorCode;
-import kr.hhplus.be.server.point.domain.entity.UserPoint;
 import kr.hhplus.be.server.point.fixture.UserPointFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
